Read Git commit hash from Debian version and disable relevant patch
Launchpad automated builds cannot get the SHA1 commit hash from the Git
history, nor it is possible to store it in the file cmake/GIT_COMMIT in
advance. However, it is possible to configure a Launchpad recipe to put
the hash at the end of the version string in debian/changelog. This
enables us to read the version from that file at build time.

Note: The top line of the Launchpad recipe must end in "-{git-commit}".
See <https://help.launchpad.net/Packaging/SourceBuilds/Recipes>.